+++ src/trx/game/enum.c
@@ -184,7 +184,6 @@ static __attribute__((constructor)) void M_Init(void)
     ENUM_MAP(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_MEDIPACKS, "remove_medipacks");
     ENUM_MAP(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_FLARES, "remove_flares");
     ENUM_MAP(GF_SEQUENCE_EVENT_TYPE, GFS_DISABLE_FLOOR, "disable_floor");
-    ENUM_MAP(GF_SEQUENCE_EVENT_TYPE, GFS_SETUP_BACON_LARA, "setup_bacon_lara");
     ENUM_MAP(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_SCIONS, "remove_scions");
     ENUM_MAP(
         GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_BINOCULARS, "remove_binoculars");