+++ src/trx/game/objects/creatures/bacon_lara.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-bool BaconLara_InitialiseAnchor(int32_t room_index);